+++ packages/react-components/react-headless-components-preview/library/bundle-size/AllComponents.fixture.js
@@ -1,6 +1,68 @@
-import * as rhc from '@fluentui/react-headless-components-preview';
+import * as Accordion from '@fluentui/react-headless-components-preview/accordion';
+import * as Avatar from '@fluentui/react-headless-components-preview/avatar';
+import * as Badge from '@fluentui/react-headless-components-preview/badge';
+import * as Breadcrumb from '@fluentui/react-headless-components-preview/breadcrumb';
+import * as Button from '@fluentui/react-headless-components-preview/button';
+import * as Card from '@fluentui/react-headless-components-preview/card';
+import * as Checkbox from '@fluentui/react-headless-components-preview/checkbox';
+import * as Dialog from '@fluentui/react-headless-components-preview/dialog';
+import * as Divider from '@fluentui/react-headless-components-preview/divider';
+import * as Drawer from '@fluentui/react-headless-components-preview/drawer';
+import * as Field from '@fluentui/react-headless-components-preview/field';
+import * as Input from '@fluentui/react-headless-components-preview/input';
+import * as Link from '@fluentui/react-headless-components-preview/link';
+import * as MessageBar from '@fluentui/react-headless-components-preview/message-bar';
+import * as ProgressBar from '@fluentui/react-headless-components-preview/progress-bar';
+import * as Popover from '@fluentui/react-headless-components-preview/popover';
+import * as Provider from '@fluentui/react-headless-components-preview/provider';
+import * as RadioGroup from '@fluentui/react-headless-components-preview/radio-group';
+import * as RatingDisplay from '@fluentui/react-headless-components-preview/rating-display';
+import * as Rating from '@fluentui/react-headless-components-preview/rating';
+import * as SearchBox from '@fluentui/react-headless-components-preview/search-box';
+import * as Select from '@fluentui/react-headless-components-preview/select';
+import * as Skeleton from '@fluentui/react-headless-components-preview/skeleton';
+import * as Slider from '@fluentui/react-headless-components-preview/slider';
+import * as SpinButton from '@fluentui/react-headless-components-preview/spin-button';
+import * as Spinner from '@fluentui/react-headless-components-preview/spinner';
+import * as Switch from '@fluentui/react-headless-components-preview/switch';
+import * as TabList from '@fluentui/react-headless-components-preview/tab-list';
+import * as Textarea from '@fluentui/react-headless-components-preview/textarea';
+import * as ToggleButton from '@fluentui/react-headless-components-preview/toggle-button';
+import * as Toolbar from '@fluentui/react-headless-components-preview/toolbar';
 
-console.log(rhc);
+console.log({
+  Accordion,
+  Avatar,
+  Badge,
+  Breadcrumb,
+  Button,
+  Card,
+  Checkbox,
+  Dialog,
+  Divider,
+  Drawer,
+  Field,
+  Input,
+  Link,
+  MessageBar,
+  ProgressBar,
+  Popover,
+  Provider,
+  RadioGroup,
+  RatingDisplay,
+  Rating,
+  SearchBox,
+  Select,
+  Skeleton,
+  Slider,
+  SpinButton,
+  Spinner,
+  Switch,
+  TabList,
+  Textarea,
+  ToggleButton,
+  Toolbar,
+});
 
 export default {
   name: 'react-headless-components-preview: entire library',