+++ packages/react-components/react-headless-components-preview/stories/src/Checkbox/index.stories.tsx
@@ -1,4 +1,4 @@
-import { Checkbox } from '@fluentui/react-headless-components-preview';
+import { Checkbox } from '@fluentui/react-headless-components-preview/checkbox';
 
 import descriptionMd from './CheckboxDescription.md';