+++ packages/react-components/react-headless-components-preview/stories/src/Dialog/DialogDefault.stories.tsx
@@ -6,7 +6,7 @@ import {
   DialogSurface,
   DialogTitle,
   DialogTrigger,
-} from '@fluentui/react-headless-components-preview';
+} from '@fluentui/react-headless-components-preview/dialog';
 
 export const Default = (): React.ReactNode => (
   <Dialog>