+++ packages/react-components/react-headless-components-preview/stories/src/Dialog/DialogNonModal.stories.tsx
@@ -6,7 +6,7 @@ import {
   DialogSurface,
   DialogTitle,
   DialogTrigger,
-} from '@fluentui/react-headless-components-preview';
+} from '@fluentui/react-headless-components-preview/dialog';
 
 /**
  * A non-modal dialog does not dim the background and does not trap focus.