+++ packages/react-components/react-headless-components-preview/stories/src/Field/index.stories.tsx
@@ -1,4 +1,4 @@
-import { Field } from '@fluentui/react-headless-components-preview';
+import { Field } from '@fluentui/react-headless-components-preview/field';
 
 import descriptionMd from './FieldDescription.md';