+++ packages/react-components/react-headless-components-preview/stories/src/Rating/index.stories.tsx
@@ -1,4 +1,4 @@
-import { Rating, RatingItem } from '@fluentui/react-headless-components-preview';
+import { Rating, RatingItem } from '@fluentui/react-headless-components-preview/rating';
 
 import descriptionMd from './RatingDescription.md';