+++ packages/react-components/react-headless-components-preview/stories/src/Select/index.stories.tsx
@@ -1,4 +1,4 @@
-import { Select } from '@fluentui/react-headless-components-preview';
+import { Select } from '@fluentui/react-headless-components-preview/select';
 
 import descriptionMd from './SelectDescription.md';