+++ packages/react-components/react-headless-components-preview/stories/src/Spinner/index.stories.tsx
@@ -1,4 +1,4 @@
-import { Spinner } from '@fluentui/react-headless-components-preview';
+import { Spinner } from '@fluentui/react-headless-components-preview/spinner';
 
 import descriptionMd from './SpinnerDescription.md';