+++ packages/react-components/react-headless-components-preview/stories/src/ToggleButton/ToggleButtonDefault.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { ToggleButton } from '@fluentui/react-headless-components-preview';
+import { ToggleButton } from '@fluentui/react-headless-components-preview/toggle-button';
 
 export const Default = (): React.ReactNode => {
   const [checked, setChecked] = React.useState(false);